services/melo/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.8.0
attrs==24.3.0
audioread==3.0.1
av==10.0.0
babel==2.16.0
blinker==1.9.0
boto3==1.36.2
botocore==1.36.2
cached-path==1.6.7
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cn2an==0.5.22
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
ctranslate2==3.24.0
cycler==0.12.1
cython==3.0.11
dateparser==1.1.8
decorator==5.1.1
deprecated==1.2.15
distance==0.1.3
docopt==0.6.2
dtw-python==1.4.4
eng-to-ipa==0.0.2
exceptiongroup==1.2.2
fastapi==0.115.6
faster-whisper==0.9.0
ffmpy==0.5.0
filelock==3.13.1
flask==3.1.0
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.2.0
fugashi==1.3.0
g2p-en==2.1.0
g2pkk==0.1.2
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gradio==3.48.0
gradio-client==0.6.1
grpcio==1.69.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib-resources==6.5.2
inflect==7.0.0
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langid==1.1.6
librosa==0.9.1
llvmlite==0.43.0
loguru==0.7.2
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mecab-python3==1.0.9
melotts @ git+https://github.com/myshell-ai/MeloTTS.git@209145371cff8fc3bd60d7be902ea69cbdb7965a
more-itertools==10.6.0
mpmath==1.3.0
myshell-openvoice @ git+https://github.com/myshell-ai/OpenVoice.git@bb79fa78a5a7a7a3d7602b7f6d48705213a039c7
narwhals==1.22.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.12
numba==0.60.0
numpy==1.22.0
onnxruntime==1.20.1
openai-whisper==20240930
orjson==3.10.14
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pip==22.2.1
plac==1.4.3
platformdirs==4.3.6
pooch==1.8.2
proces==0.1.7
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic-core==2.27.2
pydub==0.25.1
pygments==2.19.1
pykakasi==2.2.1
pyparsing==3.2.1
pypinyin==0.50.0
pyreadline3==3.5.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
pyyaml==6.0.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.11.1
scikit-learn==1.6.1
scipy==1.11.4
semantic-version==2.10.0
setuptools==63.2.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.0
starlette==0.41.3
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.13.3
torch==2.1.2
torchaudio==2.1.2
tqdm==4.67.1
transformers==4.27.4
txtsplit==1.0.0
typing-extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
unidecode==1.3.7
unidic==1.1.0
unidic-lite==1.0.8
urllib3==2.3.0
uv==0.5.21
uvicorn==0.34.0
wasabi==0.10.1
wavmark==0.0.3
websockets==11.0.3
werkzeug==3.1.3
whisper-timestamped==1.14.2
win32-setctime==1.2.0
wrapt==1.17.2
